\name{odrcheck-package}
\alias{odrcheck-package}
\alias{odrcheck}
\docType{package}
\title{\packageTitle{odrcheck}}
\description{\packageDescription{odrcheck}}
\section{Package Content}{\packageIndices{odrcheck}}
\author{\packageAuthor{odrcheck}}
\section{Maintainer}{\packageMaintainer{odrcheck}}
\keyword{package}
